Rename privacy module to extension
