+++ app/build.gradle
@@ -179,7 +179,7 @@ if (project.hasProperty("telemetry") && project.property("telemetry") == "true")
 dependencies {
     implementation project(':freshtab')
     implementation project(':dat')
-    implementation project(':privacy')
+    implementation project(':extension')
 
     implementation Deps.mozilla_concept_engine
     implementation Deps.mozilla_concept_tabstray
