+++ extension/src/main/res/values/strings.xml
